hr-assistant-pro/medical-ontology-mcp/requirements.txt
# Core dependencies
mcp>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
polars>=0.20.0  # For efficient large file processing

# Async support
aiohttp>=3.9.0
aiofiles>=23.2.0

# Search and indexing
whoosh>=2.7.4  # Pure Python full-text search
pyahocorasick>=2.0.0  # Fast string matching

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=24.1.0
python-json-logger>=2.0.7

# Optional: Database support
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0  # For PostgreSQL
# motor>=3.3.0  # For MongoDB

# Optional: Caching
# redis>=5.0.0
# aiocache>=0.12.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0

# Research/Data Science (optional)
# jupyter>=1.0.0
# matplotlib>=3.8.0
# seaborn>=0.13.0
# scikit-learn>=1.3.0